requirements.txt
asgiref==3.7.2
attrs==23.1.0
Brotli==1.0.9
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
psycopg2==2.9.6
pycodestyle==2.10.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.10
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
